testing with multiple files in coverageData
