+++ .github/workflows/build.yml
@@ -27,6 +27,7 @@ jobs:
       - name: Upload coverage
         uses: ./
         with:
+          coverageData: ['coverage/lcov.info','packages/package-1/coverage/lcov.info','packages/package-2/coverage/lcov.info']
           coverageEndpoint: http://e3df-75-159-190-236.ngrok.io/api/coverage
           coverageToken: ${{ secrets.COVERAGE_TOKEN }}
         env:
