+++ src/send-data.ts
@@ -6,6 +6,7 @@ const sendData = async (url: string, prData: PRData): Promise<PRData> => {
   const postData: CoverageJson = {
     id: prData.repositoryId,
     baseRef: prData.baseRef,
+    ref: prData.ref,
     hash: prData.sha,
     actor: prData.actor,
     linesHit: prData.coverage.lines.hit,
@@ -25,15 +26,11 @@ const sendData = async (url: string, prData: PRData): Promise<PRData> => {
     const response = await axios.post(url, postData);
 
     if(response.headers.responseType === 'difference'){
-      console.log('response type was difference ')
       prData.coverage.lines.diff = response.data.linesDifference;
       prData.coverage.functions.diff = response.data.functionsDifference;
       prData.coverage.branches.diff = response.data.branchesDifference;
 
-
-
     } else if(response.headers.responseType === 'comment'){
-      console.log('response type was comment')
       prData.message = response.data.comment
     }
